half seperate
